requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

langgraph
langchain
langchain-openai

openai

pytest
pytest-cov
coverage

ruff
mypy
bandit

pip-audit
pipdeptree

gitpython
PyGithub
rich
